Added MedicaitonSchedulesLayout component for organization
